+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @codemod-com/backend
 
+## 0.0.159
+
+### Patch Changes
+
+- daea47d: add support for ESM
+- Updated dependencies [daea47d]
+  - @codemod-com/runner@1.2.0
+  - @codemod-com/utilities@1.1.8
+  - @codemod-com/database@1.0.0
+
 ## 0.0.158
 
 ### Patch Changes